move categorizer into categorization folder
