+++ law.cfg
@@ -35,7 +35,7 @@ default_common_version: common3
 production_modules: hbw.production.{weights,features,ml_inputs,categories,gen_hbw_decay,neutrino,synchronization}, hbw.ml.stats
 calibration_modules: columnflow.calibration.jets, hbw.calibration.default
 selection_modules: hbw.selection.{jet,common,sl_remastered,dl_remastered}
-categorization_modules: hbw.selection.categories
+categorization_modules: hbw.categorization.categories
 weight_production_modules: hbw.weight.default
 ml_modules: hbw.ml.base, hbw.ml.derived.{sl,dl,sl_res}
 inference_modules: hbw.inference.{sl,dl,sl_res}
